Modularize main.js
